requirements.txt
firecrawl-py
pandas
openpyxl
py-dotenv
groq
psaw


# ScrapegraphAI Dependencies

# scrapegraphai==0.9 
# playwright
# nest_asyncio


